+++ test_package/numpy/test.py
@@ -3,4 +3,4 @@ import sys
 import numpy as np
 
 
-sys.exit(not np.test(verbose=2))
+sys.exit(not np.test(verbose=2, extra_argv=["-n", "auto"]))