+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## v1.6.0 | In development
 
+- Added tests for Python 3.11 and set minimum supported Python version to 3.9.8.
 - Recipe build performance has been improved significantly and Conan cache usage has been reduced: 
   - In cases where only the `packages` option changes, the recipe no longer requires CPython to be re-compiled (Linux, macOS) or re-downloaded (Windows) every single time. Instead, we take advantage of the Conan cache: the new `embedded_python-core` package contains all baseline binaries (without any `pip` packages). `embedded_python` builds on top of `-core` by adding the `pip` packages and can reuse any compatible `-core` package from the cache.
   - The Python packages are now installed directly into the `package` folder instead of going via the `build` folder. This speeds up the packaging and reduces space usage since there's no more file duplication.